ECommerce/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.9
itsdangerous==2.0.1
mysql-connector-python==8.0.28